ajuste valores monetarios

+++ api/app/src/database/migrations/20240207222335_createProducts.js
@@ -9,7 +9,7 @@ exports.up = function(knex) {
         table.text("description").notNullable()
         table.enum("category", ['brasileira','vegana','vegetariana', 'salada','japonesa','chinesa','mexicana','nordestina','hambúrguer',
         'pizza','massas','álcool','bebidas','refrigerantes','gourmet','almoço','italiana','tortas','salgados' ,'doces','grelhados','sanduíche','outros']).notNullable()
-        table.double("price").notNullable()
+        table.decimal("price").notNullable()
         table.text("avatar").notNullable()
         table.timestamp('created_at').defaultTo(knex.raw('CURRENT_TIMESTAMP(0)'))
         table.timestamp('updated_at').defaultTo(knex.raw('CURRENT_TIMESTAMP(0)'))
