+++ api/app/src/database/migrations/20240207222401_createCarts.js
@@ -6,7 +6,7 @@ exports.up = function(knex) {
     return knex.schema.createTable("carts", (table) => {
         table.uuid("id", { primaryKey: true });
         table.uuid("user_id").references('id').inTable('users').onDelete('CASCADE').notNullable
-        table.double("amount").defaultTo(0)
+        table.decimal("amount").defaultTo(0)
         table.timestamp('created_at').defaultTo(knex.raw('CURRENT_TIMESTAMP(0)'))
         table.timestamp('updated_at').defaultTo(knex.raw('CURRENT_TIMESTAMP(0)'))
     })
