+++ api/app/src/database/migrations/20240214132723_createPayments.js
@@ -6,7 +6,7 @@ exports.up = function (knex) {
   return knex.schema.createTable("payments", (table) => {
     table.uuid("id", { primaryKey: true });
     table.enum("method", ["Cartâo de Crédito", "Pix"]);
-    table.double("amount").notNullable();
+    table.decimal("amount").notNullable();
     table.uuid("cart_id").references("id").inTable("carts").notNullable();
     table.timestamp("date").defaultTo(knex.raw("CURRENT_TIMESTAMP(0)"));
   });
